+++ data/examples/module-header/preceding-comment-with-haddock-out.hs
@@ -1,0 +1,11 @@
+{-
+   Here we go.
+-}
+
+-- | This is the module's Haddock.
+module Main
+  ( main,
+  )
+where
+
+main = return ()