+++ data/examples/module-header/preceding-comment-with-haddock.hs
@@ -1,0 +1,8 @@
+{-
+   Here we go.
+-}
+
+-- | This is the module's Haddock.
+module Main (main) where
+
+main = return ()