+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 * Fixed a minor inconsistency between formatting of normal and foreign type
   signatures. See [issue 408](https://github.com/tweag/ormolu/issues/408).
 
+* Fixed a bug when comment before module header with Haddock was moved
+  inside the export list. See [issue
+  430](https://github.com/tweag/ormolu/issues/430).
+
 ## Ormolu 0.0.1.0
 
 * Initial release.